refactor(ui): factorise tile and subTile computed properties
